LangGraphPracticals/requirements.txt
pydantic
langchain
langgraph
langchain-core
langchain-community
langchain-groq
python-dotenv
arxiv
wikipedia
langchain-tavily